moved IGitRepositoryCommands into IGitRepository
